+++ python/GafferVDBUI/__init__.py
@@ -53,5 +53,6 @@ from . import PointsGridToPointsUI
 from . import SphereLevelSetUI
 from . import PointsToLevelSetUI
 from . import VolumeScatterUI
+from . import VDBInspector
 
 __import__( "IECore" ).loadConfig( "GAFFER_STARTUP_PATHS", subdirectory = "GafferVDBUI" )