+++ Changes.md
@@ -1,10 +1,26 @@
 1.6.x.x (relative to 1.6.0.0a1)
 =======
 
+Features
+--------
+
+- SceneInspector :
+  - Rewrote using the same framework as LightEditor, AttributeEditor and RenderPassEditor.
+  - Added EditScope support, with in-place editing of attributes, options and shader parameters. Editing for other properties is planned for future versions.
+  - Improved comparison abilities, including the added ability to compare two render passes.
+  - Added subsections for options and attributes, splitting them by category.
+  - Added column resizing, including automatic resizing to show long attribute names.
+  - Added inspection of all shaders and parameters in shader networks.
+  - Added more detailed inspection of meshes and curves.
+  - Added more detailed inspection of primitive variables.
+
 Improvements
 ------------
 
 - Arnold : Improved readability of shader node names as they appear in statistics and `.ass` files. They are now formatted as `shader:{handle}:{uniqueId}` where `{handle}` is the Gaffer node name of the assigned shader, or if loaded from USD, the prim name.
+- PrimitiveInspector : Added `location` widget for selecting the location to be inspected.
+- OptionalValuePlugValueWidget : The widget for the `enabled` plug can now be customised using `plugValueWidget:type` metadata.
+- PathListingWidget : Improved formatting of TypedVectorData.
 
 Fixes
 -----
@@ -17,10 +33,20 @@ Fixes
   - Fixed unwanted vertical scrolling when switching tabs.
   - Fixed flickering when switching tabs.
 
+API
+---
+
+- Path : Added `contextProperty()` method.
+
 Breaking Changes
 ----------------
 
 - Arnold : Changed the naming of shader nodes in the generated Arnold scene.
+- Path : Removed `inspectionContext()` method. Use `contextProperty( "inspector:context" )` instead.
+- SceneInspector :
+  - Removed `setTargetPaths()` and `getTargetPaths()` methods. Use the `location` and `compare.location` settings plugs instead.
+  - Removed `registerSection()` method.
+  - Removed `Diff`, `SideBySideDiff`, `TextDiff`, `Row`, `Inspector`, `DiffRow`, `DiffColumn`, `Section`, `LocationSection`, `HistorySection` and `SetsSection` classes.
 
 [^1]: To be omitted from the notes for the final 1.6.0.0 release.